refactor to more generic availability state model
