+++ server/app/models/host_node.rb
@@ -31,7 +31,8 @@ class HostNode
   field :agent_version, type: String
   field :docker_version, type: String
   field :connected_at, type: Time
-  field :evacuated, type: Boolean, default: false
+  #field :evacuated, type: Boolean, default: false
+  field :availability, type: String, default: 'active'
 
   embeds_many :volume_drivers, class_name: 'HostNodeDriver'
   embeds_many :network_drivers, class_name: 'HostNodeDriver'
