+++ server/app/services/grid_service_scheduler.rb
@@ -16,7 +16,7 @@ class GridServiceScheduler
   def initialize(strategy)
     @strategy = strategy
     @filters = [
-      Scheduler::Filter::Evacuated.new,
+      Scheduler::Filter::Availability.new,
       Scheduler::Filter::Ephemeral.new,
       Scheduler::Filter::VolumePlugin.new,
       Scheduler::Filter::VolumeInstance.new,
